requirements.txt
fastapi
uvicorn
python-dotenv
pydantic-settings
pymupdf
python-pptx
python-docx
reportlab
openai
groq
pinecone
upstash-redis
httpx
